refactor: rename TopNavbar to NavbarTop
